pyproject.toml
[project]
name = "trade-bot-v2"
version = "0.1.0"
description = "Algorithmic trading bot for Indian Markets using XTS API"
requires-python = ">=3.10"
dependencies = [
    "polars",
    "pytz",
    "requests",
    "python-engineio==3.13.0",
    "python-socketio==4.6.0",
    "websocket-client==0.57.0",
    "pydantic-settings",
    "pydantic",
    "pymongo",
    "typer",
    "questionary",
    "aiohttp",
    "fastapi",
    "uvicorn",
    "scikit-learn",
    "xgboost",
    "joblib",
    "numpy",
    "httpx",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "ruff"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PL",  # pylint
    "RUF", # ruff-specific rules
]
ignore = [
    "PLR0913", # Too many arguments in function definition
    "PLR2004", # Magic value used in comparison
    "E501",    # Line too long (aesthetic)
    "PLC0415", # Import outside top-level (common in this project)
    "E402",    # Module level import not at top of file
]


[tool.ruff.lint.per-file-ignores]
# Ignore naming conventions for external API wrappers and legacy logic
"packages/xts/**/*.py" = ["N801", "N802", "N803", "N806", "N815"]
"tmp_test/**/*.py" = ["N801", "N802", "N803", "N806", "N815"]
"packages/settings.py" = ["N802"]
"packages/utils/log_utils.py" = ["N802"]


# Ignore complexity and local imports in tests and maintenance scripts
"tests/**/*.py" = ["PLR0912", "PLR0915", "PLC0415", "E402"]
"scripts/**/*.py" = ["PLR0912", "PLR0915", "E402"]

[tool.ruff.lint.isort]
known-first-party = ["packages", "apps"]


[tool.uv]
package = false


